Kill shadcn and tailwindcss, use radix theme, wip
