+++ package.json
@@ -14,6 +14,7 @@
     "@deck.gl/layers": "8.8.9",
     "@mgcth/deck.gl-pmtiles": "git@github.com:mgcth/deck.gl-pmtiles.git#9ee5dcd",
     "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-navigation-menu": "^1.1.4",
     "@radix-ui/react-popover": "^1.0.7",
@@ -21,9 +22,8 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-tabs": "^1.0.4",
+    "@radix-ui/themes": "^2.0.3",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
-    "cmdk": "^0.2.0",
     "deck.gl": "8.8.9",
     "eslint-plugin-react": "^7.33.2",
     "localforage": "^1.10.0",
@@ -35,8 +35,6 @@
     "react-router-dom": "^6.21.0",
     "react-select": "^5.8.0",
     "sort-by": "^1.2.0",
-    "tailwind-merge": "^2.1.0",
-    "tailwindcss-animate": "^1.0.7",
     "zustand": "^4.4.7"
   },
   "devDependencies": {
@@ -45,6 +43,7 @@
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
+    "@vanilla-extract/vite-plugin": "^3.9.5",
     "@vitejs/plugin-react-swc": "^3.3.2",
     "autoprefixer": "^10.4.16",
     "eslint": "^8.45.0",
@@ -53,7 +52,6 @@
     "eslint-plugin-react-refresh": "^0.4.3",
     "postcss": "^8.4.32",
     "prettier": "3.1.0",
-    "tailwindcss": "^3.4.0",
     "typescript": "^5.0.2",
     "vite": "^4.4.5"
   }
