--- src/vite-env.d.ts
@@ -1,5 +1,0 @@
-/// <reference types="vite/client" />
-declare module "tailwind-config" {
-  const config: Config;
-  export default config;
-}
