+++ vite.config.ts
@@ -1,24 +1,15 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
-import path from "path";
+import { vanillaExtractPlugin } from "@vanilla-extract/vite-plugin";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   build: {
+    target: "esnext",
     commonjsOptions: {
-      include: ["tailwind-config.js", "node_modules/**"],
+      include: ["node_modules/**"],
     },
   },
 
-  optimizeDeps: {
-    include: ["tailwind-config"],
-  },
-
-  plugins: [react()],
-  resolve: {
-    alias: {
-      "@": path.resolve(__dirname, "./src"),
-      "tailwind-config": path.resolve(__dirname, "./tailwind.config.js"),
-    },
-  },
+  plugins: [react(), vanillaExtractPlugin()],
 });
